+++ src/main/java/org/intellij/sequencer/generator/KtSequenceGenerator.java
@@ -1,18 +1,25 @@
 package org.intellij.sequencer.generator;
 
 import com.intellij.openapi.diagnostic.Logger;
-import com.intellij.psi.PsiCallExpression;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiMethod;
 import com.intellij.util.containers.Stack;
+import org.apache.log4j.Level;
+import org.intellij.sequencer.Constants;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+import org.jetbrains.kotlin.descriptors.CallableDescriptor;
+import org.jetbrains.kotlin.idea.caches.resolve.ResolutionUtils;
 import org.jetbrains.kotlin.psi.*;
+import org.jetbrains.kotlin.resolve.DescriptorToSourceUtils;
+import org.jetbrains.kotlin.resolve.calls.model.ResolvedCall;
+import org.jetbrains.kotlin.resolve.lazy.BodyResolveMode;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
 public class KtSequenceGenerator extends KtVisitorVoid implements IGenerator {
-    private final Stack<PsiCallExpression> exprStack = new Stack<>();
+    private final Stack<KtExpression> exprStack = new Stack<>();
     private final Stack<CallStack> callStack = new Stack<>();
     private static final Logger LOGGER = Logger.getInstance(KtSequenceGenerator.class.getName());
 
@@ -23,15 +30,35 @@ public class KtSequenceGenerator extends KtVisitorVoid implements IGenerator {
 
     public KtSequenceGenerator(SequenceParams params) {
         this.params = params;
+        LOGGER.setLevel(Level.DEBUG);
     }
 
     @Override
     public CallStack generate(PsiElement psiElement) {
-        return generate((KtFunction) psiElement);
+        if (LOGGER.isDebugEnabled()) {
+            LOGGER.debug("[generate]" + psiElement.getText());
+        }
+        if (psiElement instanceof KtFunction) {
+            return generate((KtFunction) psiElement);
+        }
+
+        if (psiElement instanceof PsiMethod) {
+            return generate((PsiMethod) psiElement);
+        }
+
+        return topStack;
     }
 
     public CallStack generate(KtFunction ktFunction) {
         ktFunction.accept(this);
+        return topStack;
+    }
+
+    public CallStack generate(PsiMethod psiMethod) {
+        CallStack javaCall = new SequenceGenerator(params).generate(psiMethod);
+        LOGGER.debug("[JAVACall]:"+javaCall.generateText());
+        currentStack.methodCall(javaCall.getMethod());
+//        makeMethodCallExceptCurrentStackIsRecursive(javaCall.getMethod());
         return topStack;
     }
 
@@ -44,18 +71,56 @@ public class KtSequenceGenerator extends KtVisitorVoid implements IGenerator {
 
     @Override
     public void visitNamedFunction(@NotNull KtNamedFunction function) {
+        if (LOGGER.isDebugEnabled()) {
+            LOGGER.debug("[visitNamedFunction]" + function.getText());
+        }
         MethodDescription method = createMethod(function);
         if (makeMethodCallExceptCurrentStackIsRecursive(method)) return;
         super.visitNamedFunction(function);
     }
 
     @Override
+    public void visitPrimaryConstructor(@NotNull KtPrimaryConstructor constructor) {
+        if (LOGGER.isDebugEnabled()) {
+            LOGGER.debug("[visitPrimaryConstructor]" + constructor.getText());
+        }
+        MethodDescription method = createMethod(constructor);
+        if (makeMethodCallExceptCurrentStackIsRecursive(method)) return;
+        super.visitPrimaryConstructor(constructor);
+    }
+
+    @Override
+    public void visitSecondaryConstructor(@NotNull KtSecondaryConstructor constructor) {
+        if (LOGGER.isDebugEnabled()) {
+            LOGGER.debug("[visitSecondaryConstructor]" + constructor.getText());
+        }
+        MethodDescription method = createMethod(constructor);
+        if (makeMethodCallExceptCurrentStackIsRecursive(method)) return;
+        super.visitSecondaryConstructor(constructor);
+    }
+
+    @Override
     public void visitCallExpression(@NotNull KtCallExpression expression) {
-        PsiElement psiElement = expression.getReference().resolve();
-//        KtNameReferenceExpression nameReferenceExpression = (KtNameReferenceExpression) expression.getCalleeExpression();
-//        PsiElement psiElement = nameReferenceExpression.getReferencedNameElement();
-        methodCall(psiElement);
+        if (LOGGER.isDebugEnabled()) {
+            LOGGER.debug("[visitCallExpression]" + expression.getText());
+        }
+        PsiElement psiElement = resolveFunction(expression);
+        if (psiElement instanceof KtClass) {
+            currentStack.methodCall(createMethod((KtClass) psiElement));
+        } else {
+            methodCall(psiElement);
+        }
+
         super.visitCallExpression(expression);
+    }
+
+
+
+    @Nullable
+    private PsiElement resolveFunction(@NotNull KtExpression expression) {
+        ResolvedCall<? extends CallableDescriptor> resolvedCall = ResolutionUtils.resolveToCall(expression, BodyResolveMode.PARTIAL);
+        PsiElement psiElement = DescriptorToSourceUtils.descriptorToDeclaration(resolvedCall.getCandidateDescriptor());
+        return psiElement;
     }
 
     private void methodCall(PsiElement psiElement) {
@@ -71,15 +136,24 @@ public class KtSequenceGenerator extends KtVisitorVoid implements IGenerator {
             LOGGER.debug("- depth = " + depth + " method = " + psiElement.getText());
             currentStack = oldStack;
         } else {
-            currentStack.methodCall(createMethod((KtNamedFunction) psiElement));
+            currentStack.methodCall(createMethod(psiElement));
         }
+    }
+
+    private MethodDescription createMethod(PsiElement psiElement) {
+        if (psiElement instanceof KtClass) return createMethod((KtClass) psiElement);
+        else if (psiElement instanceof KtNamedFunction) return createMethod((KtNamedFunction) psiElement);
+        // todo Something else
+        return null;
     }
 
     private MethodDescription createMethod(KtNamedFunction function) {
         ParamPair paramPair = extractParameters(function.getValueParameters());
         ClassDescription classDescription;
         if (function.isTopLevel()) {
-            classDescription = ClassDescription.TOP_LEVEL_FUN;
+            String filename = function.getContainingKtFile().getName();
+            filename = filename.replace(".kt", "_kt");
+            classDescription = ClassDescription.getFileNameAsClass(filename); //ClassDescription.TOP_LEVEL_FUN;
         } else if (function.isLocal()) {
             classDescription = ClassDescription.ANONYMOUS_CLASS;
         } else {
@@ -100,6 +174,46 @@ public class KtSequenceGenerator extends KtVisitorVoid implements IGenerator {
                 classDescription,
                 attributes, function.getName(), returnType,
                 paramPair.argNames, paramPair.argTypes
+        );
+    }
+
+    private MethodDescription createMethod(KtSecondaryConstructor constructor) {
+        ParamPair paramPair = extractParameters(constructor.getValueParameters());
+        ClassDescription classDescription = new ClassDescription(constructor.getName(), new ArrayList<>());
+
+        List<String> attributes = createAttributes(constructor.getModifierList());
+        String returnType = constructor.getName();
+
+        return MethodDescription.createMethodDescription(
+                classDescription,
+                attributes, Constants.CONSTRUCTOR_METHOD_NAME, returnType,
+                paramPair.argNames, paramPair.argTypes
+        );
+    }
+
+    private MethodDescription createMethod(KtPrimaryConstructor constructor) {
+        ParamPair paramPair = extractParameters(constructor.getValueParameters());
+        ClassDescription classDescription = new ClassDescription(constructor.getName(), new ArrayList<>());
+
+        List<String> attributes = createAttributes(constructor.getModifierList());
+        String returnType = constructor.getName();
+
+        return MethodDescription.createMethodDescription(
+                classDescription,
+                attributes, Constants.CONSTRUCTOR_METHOD_NAME, returnType,
+                paramPair.argNames, paramPair.argTypes
+        );
+    }
+
+    private MethodDescription createMethod(KtClass ktClass) {
+        ClassDescription classDescription = new ClassDescription(ktClass.getFqName().asString(), new ArrayList<>());
+        List<String> attributes = createAttributes(ktClass.getModifierList());
+        String returnType = ktClass.getName();
+
+        return MethodDescription.createMethodDescription(
+                classDescription,
+                attributes, Constants.CONSTRUCTOR_METHOD_NAME, returnType,
+                new ArrayList<>(), new ArrayList<>()
         );
     }
 
@@ -125,8 +239,17 @@ public class KtSequenceGenerator extends KtVisitorVoid implements IGenerator {
      * @return String
      */
     private String getType(KtTypeReference typeReference) {
-        KtUserType type = (KtUserType) Objects.requireNonNull(typeReference).getTypeElement();
-        return Objects.requireNonNull(type).getReferencedName();
+        KtTypeElement typeElement = typeReference.getTypeElement();
+
+        if (typeElement instanceof KtNullableType) {
+            typeElement = ((KtNullableType) typeElement).getInnerType();
+        }
+
+        if (typeElement instanceof KtUserType) {
+            return ((KtUserType) typeElement).getReferencedName();
+        }
+
+        return "Unit";
     }
 
     private boolean makeMethodCallExceptCurrentStackIsRecursive(MethodDescription method) {